backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
SQLAlchemy-Utils==0.41.1
email-validator==2.1.0
Jinja2==3.1.2
schedule==1.2.0
fake-useragent==1.4.0
aiohttp==3.9.1
python-dateutil==2.8.2
pytz==2023.3
cloudscraper==1.2.71